+++ src/Streamly/Internal/Network/Inet/TCP.hs
@@ -399,7 +399,7 @@ writeWithBufferOf
     -> PortNumber
     -> Fold m Word8 ()
 writeWithBufferOf n addr port =
-    FL.chunksOf n (writeNUnsafe n) (writeChunks addr port)
+    FL.groupsOf n (writeNUnsafe n) (writeChunks addr port)
 
 -- | Write a stream to the supplied IPv4 host address and port number.
 --